Update the status prober to work with dynamic paths
Because probes on the HTTPRoute now have unique paths
we no longer use a vanilla prober from knative/networking.

The changes make the prober less coupled to Ingress and it requires
consumers to build up a list of backends URLs to probe.

The probe target lister now merges backends and URLs with their
corresponding proxy pod IPs and ports. The alternative
would be to look up HTTPRoutes in the informer cache but theres
no guarantee this is up to date.