Fork the knative/networking Status.Prober (#679)
* copy the networking status prober

* Add IsProbeActive to query probing state

* fix boilerplate

* Update the status prober to work with dynamic paths

Because probes on the HTTPRoute now have unique paths
we no longer use a vanilla prober from knative/networking.

The changes make the prober less coupled to Ingress and it requires
consumers to build up a list of backends URLs to probe.

The probe target lister now merges backends and URLs with their
corresponding proxy pod IPs and ports. The alternative
would be to look up HTTPRoutes in the informer cache but theres
no guarantee this is up to date.

* Don't override the desired Probe path

* add test coverage for Backends.AddURL

* Add test coverage for IsProbeActive - change function signature

* address some linter feedback

* run ./hack/update-deps.sh

* Drop anonymous function in DoProbe & use RWLock

* add godoc

* fix test error messages and comments

* rename state to ingState

* Add back TestCancelIngressProbing

* linter: use switch statement

* use test logger so log output from DoProbes is tied to the test

* fix error message