Fix probing getting stuck forever (#698)
A stale httproute in the informer cache can result in probing to get stuck.

This would cause the reconciler to detect a new backend and revert from
our 'transitioning' phase back to the endpoint phase.

The changes tweaks the state machine so that we only consider 'new'
backends if the ingress hash has changed.